requirements.txt
uuid
python-dotenv


